requirements.txt
streamlit-chat
python-dotenv
langchain
openai
langchain-openai